+++ backend/footprint/src/test/java/foot/footprint/service/commentLike/CommentLikeServiceTest.java
@@ -9,9 +9,9 @@ import static org.mockito.Mockito.verify;
 import foot.footprint.domain.article.dao.FindArticleRepository;
 import foot.footprint.domain.article.domain.Article;
 import foot.footprint.domain.article.exception.NotMatchMemberException;
-import foot.footprint.domain.commentLike.application.ChangeGroupedCommentLike;
-import foot.footprint.domain.commentLike.application.ChangePrivateCommentLike;
-import foot.footprint.domain.commentLike.application.ChangePublicCommentLike;
+import foot.footprint.domain.commentLike.application.ChangeGroupedCommentLikeService;
+import foot.footprint.domain.commentLike.application.ChangePrivateCommentLikeService;
+import foot.footprint.domain.commentLike.application.ChangePublicCommentLikeService;
 import foot.footprint.domain.commentLike.dao.CommentLikeRepository;
 import foot.footprint.domain.group.dao.ArticleGroupRepository;
 import foot.footprint.global.error.exception.NotAuthorizedOrExistException;
@@ -38,13 +38,13 @@ public class CommentLikeServiceTest {
     private ArticleGroupRepository articleGroupRepository;
 
     @InjectMocks
-    private ChangePublicCommentLike changePublicCommentLike;
+    private ChangePublicCommentLikeService changePublicCommentLikeService;
 
     @InjectMocks
-    private ChangePrivateCommentLike changePrivateCommentLike;
+    private ChangePrivateCommentLikeService changePrivateCommentLikeService;
 
     @InjectMocks
-    private ChangeGroupedCommentLike changeGroupedCommentLike;
+    private ChangeGroupedCommentLikeService changeGroupedCommentLikeService;
 
     @Test
     @DisplayName("댓글 좋아요 변화 - 전체지도 시")
@@ -56,7 +56,7 @@ public class CommentLikeServiceTest {
         given(commentLikeRepository.saveCommentLike(any())).willReturn(1);
 
         //when
-        changePublicCommentLike.changeMyLike(1L, 1L, false, 1L);
+        changePublicCommentLikeService.changeMyLike(1L, 1L, false, 1L);
 
         //then
         verify(commentLikeRepository, times(1)).saveCommentLike(any());
@@ -66,7 +66,7 @@ public class CommentLikeServiceTest {
         given(commentLikeRepository.deleteCommentLike(any(), any())).willReturn(1);
 
         //when
-        changePublicCommentLike.changeMyLike(1L, 1L, true, 1L);
+        changePublicCommentLikeService.changeMyLike(1L, 1L, true, 1L);
 
         //then
         verify(commentLikeRepository, times(1)).deleteCommentLike(any(), any());
@@ -76,7 +76,7 @@ public class CommentLikeServiceTest {
 
         //when & then
         assertThatThrownBy(
-            () -> changePublicCommentLike.changeMyLike(1L, 1L, true, 1L))
+            () -> changePublicCommentLikeService.changeMyLike(1L, 1L, true, 1L))
             .isInstanceOf(NotExistsException.class);
     }
 
@@ -90,7 +90,7 @@ public class CommentLikeServiceTest {
 
         //when & then
         assertThatThrownBy(
-            () -> changePrivateCommentLike.changeMyLike(1L, 1L, true, 2L))
+            () -> changePrivateCommentLikeService.changeMyLike(1L, 1L, true, 2L))
             .isInstanceOf(NotMatchMemberException.class);
     }
 
@@ -105,7 +105,7 @@ public class CommentLikeServiceTest {
 
         //when & then
         assertThatThrownBy(
-            () -> changeGroupedCommentLike.changeMyLike(1L, 1L, true, 2L))
+            () -> changeGroupedCommentLikeService.changeMyLike(1L, 1L, true, 2L))
             .isInstanceOf(NotAuthorizedOrExistException.class);
     }
